requirements.txt
opencv-python
numpy
matplotlib
streamlit
ffmpeg-python
